Added automatic datafile generation
